#16 refactor: 일관성있는 컨벤션을 위해 type을 types.ts 파일로 이동
