+++ src/features/stock/api/index.ts
@@ -1,28 +1,6 @@
 import { fetcherWithoutCredentials } from "@/api/fetcher";
 import { Response } from "@/api/types";
-
-export type StockSearchItem = {
-  stockCode: string;
-  tickerSymbol: string;
-  companyName: string;
-  companyNameEng: string;
-  market: string;
-};
-
-export type StockItem = {
-  stockCode: string;
-  tickerSymbol: string;
-  companyName: string;
-  companyNameEng: string;
-  market: string;
-  currentPrice: number;
-  dailyChange: number;
-  dailyChangeRate: number;
-  sector: string;
-  annualDividend: number;
-  annualDividendYield: number;
-  dividendMonths: number[];
-};
+import { StockItem, StockSearchItem } from "./types";
 
 export const postStockSearch = async (query: string) => {
   const res = await fetcherWithoutCredentials.post<Response<StockSearchItem[]>>(
