+++ src/features/stock/api/types.ts
@@ -1,0 +1,22 @@
+export type StockSearchItem = {
+  stockCode: string;
+  tickerSymbol: string;
+  companyName: string;
+  companyNameEng: string;
+  market: string;
+};
+
+export type StockItem = {
+  stockCode: string;
+  tickerSymbol: string;
+  companyName: string;
+  companyNameEng: string;
+  market: string;
+  currentPrice: number;
+  dailyChange: number;
+  dailyChangeRate: number;
+  sector: string;
+  annualDividend: number;
+  annualDividendYield: number;
+  dividendMonths: number[];
+};
